fix: resolve CI failures and merge conflicts
