+++ components/features/employees/OnboardingBadge.tsx
@@ -1,4 +1,4 @@
-import Badge from "@/components/ui/Badge";
+import { Badge } from "@/components/ui/badge";
 import type { OnboardingStatus } from "@/types";
 import { CheckCircle2, CircleDashed, Circle } from "lucide-react";
