--- components/ui/Badge.tsx
@@ -1,25 +1,0 @@
-import React from "react";
-
-interface BadgeProps {
-  children: React.ReactNode;
-  variant?: "success" | "warning" | "error" | "info" | "default";
-  className?: string;
-}
-
-const variants = {
-  success: "bg-green-100 text-green-800",
-  warning: "bg-yellow-100 text-yellow-800",
-  error: "bg-red-100 text-red-800",
-  info: "bg-blue-100 text-blue-800",
-  default: "bg-gray-100 text-gray-700",
-};
-
-export default function Badge({ children, variant = "default", className = "" }: BadgeProps) {
-  return (
-    <span
-      className={`inline-flex items-center px-2 py-0.5 rounded-full text-xs font-medium transition-colors ${variants[variant]} ${className}`}
-    >
-      {children}
-    </span>
-  );
-}
